+++ lib/domain/services/pdf_frontend_prescan_manager.dart
@@ -270,6 +270,61 @@ Future<Response> loadPdftoApi({
       
     }
 
+    Future<Response> getTopicContextFromApi({
+      required String itemName,
+      required String inputIdentification,
+  
+    }) async {
+
+       try {
+      debug('Reading next auth nounce from local storage...');
+      final nounce = await readNextAuthNounce();
+      final hintCookies = readCookie('hint');      
+
+      if (nounce != null) {
+        defaultHeadersPdfApi['T-A-N'] = nounce;
+        
+      }
+
+      if (hintCookies != null) {
+        defaultHeadersPdfApi['A-A-N'] = hintCookies;
+      }
+
+    } catch (e) {
+      debug('Error reading next auth nounce: $e');
+    }
+
+
+      final adapter = ApiAdapter(
+        defaultHeaders: {
+          ...defaultHeadersPdfApi,
+        
+        },
+      );
+
+      // /context/{item_name}/{input_identification}
+
+      final url = Uri.parse('$baseUrl/context/$itemName/$inputIdentification}');
+
+      final response = await adapter.get(url);
+
+      if (response.statusCode == 200) {
+      final headers = response.headers;
+      // get next auth nounce key = 'n-a-n'
+      final nan = headers['n-a-n'];
+      if (nan != null) {
+        await saveNextAuthNounce(nan);
+        debug('Next auth nounce updated: $nan from getCards response headers.');
+      } else {
+        warning('No next auth nounce found in response headers.');
+
+      }
+    }
+
+    return response;
+      
+    }
+