+++ CHANGELOG.md
@@ -51,6 +51,7 @@ All notable changes to this project will be documented in this file.
 - *(cards)* Added a request btn for user request new cards
 - *(cards)* Added a success msg after new app solicitation
 - *(app)* Added desktop configuration screen
+- *(topics)* Added redirection and loading data from context for topic
 
 ### 🐛 Bug Fixes